[Silver III] Title: 주유소, Time: 368 ms, Memory: 36504 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -10,25 +10,25 @@ public class Main {
         StringTokenizer st;
 
         int N = Integer.parseInt(br.readLine()); // 도시의 수
-        int[] gasPrice = new int[N]; // 도시별 기름값
-        int[] distance = new int[N-1]; // 다음도시간 거리
+        long[] gasPrice = new long[N]; // 도시별 기름값
+        long[] distance = new long[N-1]; // 다음도시간 거리
 
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i < N - 1; i++) {
-            int d = Integer.parseInt(st.nextToken());
+            long d = Integer.parseInt(st.nextToken());
             distance[i] = d;
         }
 
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i < N; i++) {
-            int price = Integer.parseInt(st.nextToken());
+            long price = Integer.parseInt(st.nextToken());
             gasPrice[i] = price;
         }
 
         br.close();
 
-        int totalCost = 0; // 총 기름값
-        int minCost = gasPrice[0]; // 기름가격 최소값
+        long totalCost = 0; // 총 기름값
+        long minCost = gasPrice[0]; // 기름가격 최소값
 
         for(int i = 0; i < N - 1; i++) {
             if(minCost > gasPrice[i]) { // 다음 방문한 도시의 기름이 더 싸면 초기화
